Add a first draft of a DIR interface, that implements opendir(), closedir(), readdir(), and the underlying getdents() function that does the actual syscall by using the dirent64 structure.
Add listdir() that uses the interface.

Add a netm library that uses listdir().

+++ src/y/pkg/libc/libc/std/int.h
@@ -137,13 +137,13 @@ typedef unsigned long long int uint_fast64_t;
 
 #if __WORDSIZE == 64
 
-typedef long long int          int64_t;
-typedef unsigned long long int uint64_t;
+typedef signed long int   int64_t;
+typedef unsigned long int uint64_t;
 
 #else
 
-typedef long int               int64_t;
-typedef unsigned long int      uint64_t;
+typedef signed long long int int64_t;
+typedef unsigned long long int uint64_t;
 
 #endif
